+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Pending
+
+### Added
+
+- Adding person-dependant `datetime_ref` to `plot_age_pyramid`
+
 ## v0.1.2 (2022-12-05)
 
 ### Added